Fix Py3 Set() Order no Longer Being Deterministic
* Programs that relied implicitly on set() order for reproducibility
across runs in Py2 can no longer expect that behaviour in Py3, which
now enables Hash-randomization by default. A simple fix is to replace
the native Py3 set() with an ordered-set implementation -- see link
below:
https://stackoverflow.com/questions/14956313/why-is-dictionary-ordering-non-deterministic
Note that an environment variable `PYTHONHASHSEED` can be set to
enable/disable this Hash-randomization, but unfortunately, it seems, one
can only set it from the terminal; i.e. it cannot be modified
programmatically.