Comments, fix handling > max_dist
We return nan instead of 0 from distance functions if beyond max
distance, because a sample exactly on the grid point will have a dist
of 0, and the weighting will be exp(0) == 1. After calculating exp,
set NaNs to 0 so samples beyond max distance have a weight of 0, rather
than NaN which will propagate.