Removing MPI-parallel Catalogue Ingestion
MPI-Allgatherv is prone to data corruption when tens of processors are
used to gather arrays on all ranks amounting to hundreds of MB of data.
Catalogue ingestion now takes place on rank 0 and events and arrivals
are saved to disk -- subsequently, all other ranks simply load these
events and arrivals from disk.